+++ vault_core/src/lib.rs
@@ -8,3 +8,5 @@ pub mod vault_ncn_ticket;
 pub mod vault_operator_delegation;
 pub mod vault_staker_withdrawal_ticket;
 pub mod vault_update_state_tracker;
+
+pub const MAX_FEE_BPS: u16 = 10_000;
